fix(code-examples): update lookup code examples to all use the `searchAsync` event
