+++ apps/code-examples/src/app/code-examples/lookup/lookup/add-item/demo.component.html
@@ -15,10 +15,10 @@
       formControlName="favoriteNames"
       idProperty="name"
       [enableShowMore]="true"
-      (searchAsync)="searchAsync($event)"
       [showAddButton]="true"
       [showMoreConfig]="showMoreConfig"
       (addClick)="addClick($event)"
+      (searchAsync)="searchAsync($event)"
     />
   </sky-input-box>
   <div
