--- apps/code-examples/src/app/code-examples/lookup/lookup/async/demo.component.html
@@ -1,36 +1,0 @@
-<form
-  class="sky-padding-even-md"
-  novalidate
-  [formGroup]="favoritesForm"
-  (ngSubmit)="onSubmit()"
->
-  <sky-input-box
-    data-sky-id="favorite-names-field"
-    helpPopoverContent="A name can be selected from the list or added as a new name."
-    hintText="Choose a name from the list or add a new name."
-    labelText="Favorite name(s)"
-    stacked
-  >
-    <sky-lookup
-      formControlName="favoriteNames"
-      idProperty="name"
-      [enableShowMore]="true"
-      (searchAsync)="searchAsync($event)"
-      [showAddButton]="true"
-      [showMoreConfig]="showMoreConfig"
-      (addClick)="addClick($event)"
-    />
-    <sky-form-error
-      *ngIf="favoritesForm.controls.favoriteNames.errors?.['letterE']"
-      errorName="letterE"
-      errorText="Please only choose names that contain the letter E"
-    />
-  </sky-input-box>
-  <div
-    class="sky-border-dark sky-rounded-corners sky-padding-even-md sky-margin-stacked-lg"
-  >
-    <strong>Form model:</strong>
-    <pre>{{ favoritesForm.value | json }}</pre>
-  </div>
-  <button class="sky-btn sky-btn-default" type="submit">Submit</button>
-</form>
