--- apps/code-examples/src/app/code-examples/lookup/lookup/async/person.ts
@@ -1,3 +1,0 @@
-export interface Person {
-  name: string;
-}
