+++ apps/code-examples/src/app/code-examples/lookup/lookup/custom-picker/picker-modal.component.scss
@@ -1,0 +1,3 @@
+.wait-container {
+  height: 100px;
+}
