+++ apps/code-examples/src/app/code-examples/lookup/lookup/custom-picker/search-results.ts
