+++ apps/code-examples/src/app/code-examples/lookup/lookup/multi-select/demo.component.html
@@ -6,20 +6,26 @@
 >
   <sky-input-box
     data-sky-id="favorite-names-field"
-    labelText="Favorite names"
-    hintText="Choose a name from the list or add a new name."
     helpPopoverContent="A name can be selected from the list or added as a new name."
+    hintText="Choose a name from the list or add a new name."
+    labelText="Favorite name(s)"
     stacked
   >
     <sky-lookup
       formControlName="favoriteNames"
-      (addClick)="onAddButtonClicked()"
-      [data]="people"
+      idProperty="name"
       [enableShowMore]="true"
-      [searchFilters]="searchFilters"
       [showAddButton]="true"
       [showMoreConfig]="showMoreConfig"
+      (searchAsync)="searchAsync($event)"
+      (addClick)="addClick($event)"
     />
+    @if (favoritesForm.controls.favoriteNames.errors?.['letterE']) {
+      <sky-form-error
+        errorName="letterE"
+        errorText="Please only choose names that contain the letter E"
+      />
+    }
   </sky-input-box>
   <div
     class="sky-border-dark sky-rounded-corners sky-padding-even-md sky-margin-stacked-lg"
