+++ apps/code-examples/src/app/code-examples/lookup/lookup/single-select/search-results.ts
@@ -1,0 +1,7 @@
+import { Person } from './person';
+
+export interface LookupAsyncDemoSearchResults {
+  hasMore: boolean;
+  people: Person[];
+  totalCount: number;
+}
