+++ apps/code-examples/src/app/home/home.component.html
@@ -490,9 +490,8 @@
         Lookup
         <ul>
           <li>
-            <a routerLink="lookup/lookup/add-item">Async with add button</a>
+            <a routerLink="lookup/lookup/add-item">Add button</a>
           </li>
-          <li><a routerLink="lookup/lookup/async">Async</a></li>
           <li>
             <a routerLink="lookup/lookup/custom-picker">Custom picker</a>
           </li>
